DRYed tests: moved measure tests to MaskTest
